Fixes for installation on OSX.
NOTE: XQuartz has to be installed for this to work!

A patch for the configure file changes the libpng suffix (from .so to .dynlib).

The build script checks for if we are using OSX (darwin), and adjusts
the X11 path if we are. There's currently no check to ensure XQuartz
is actually installed.